Skeleton for the rust-mosaic binary
I created a simple binary, library, and module.

I created a blockchain module in order to wrap the web3 dependency.
That way we can easily replace it or have a breaking upgrade if it is
required in the future. Also the rest of the code does note need to
know about the specifics of web3.

I added simple example tests.

Configuration options are loaded from the environment variables as
per [12 Factor App](https://12factor.net/).

Added a config for Travis CI and added the GitHub repository to
Travis CI to build PRs and `develop` daily.