Added example of how to use web3
Abstracted away from ethereum so that all web3 functionality should be
within a single file (ethereum.rs).

In this simple example, all the accounts from the source chain are
printed to std out.

Error handling should definitely be improved.
There are no unit tests yet, as there is no meaningful functionality
yet. It is important that testing is done well once concrete use cases
get implemented and that we do not need to make actual calls to an
ethereum test net in order to run the mosaic tests.