Added address parsing for blockchain addresses
Updated the address to include a method to create it from a string that
is a hex representation of the 20 bytes.

Added an error type to handle any kind of error that can occur as a
result of interaction with the blockchain module. The only kind so far
is an error that an address could not be parsed.

Please note the open TODOs. Someone with more rust experience should
give feedback as to what the canonical/proper way is to handle these
things.