Dockerfile and docker compose
Dockerfile to build a rust-mosaic image. It is a multi-stage build based
on debian. I chose debian as the rust image is based on debian and rust
binaries use dynamically linked libraries. We could compile the binary
with musl to use a slimmer base image for the binary distribution.

Docker compose file to run rum alongside two geth nodes: rinkeby and
test utility chain 1409.

The `rpcaddr` and `rpcvhosts` parameters in the compose file should not
allow any connection from anywhere. I will create an issue to fix that.
I will also create an issue to enable the loading of the chain data into
the containers so that not everything must be synced from the start.