Rum now reads validator addresses from ENV
Added new ENV variables to the config so that it reads both addresses
for origin and auxiliary validators from the environment.

As an example I also added a method to sign data and use it when running
the rust node. The execution is just an example and would be replaced by
actual validator tasks.

There is no good error handling yet. It may be the case that we want to
retry a couple of times before giving up.

Right now I made it so that the user must enter the password when
starting the node so that rum can unlock the addresses on the
ethereum node. I cannot think of a better way, but this cannot be
automated nicely. Does anyone have an idea? Adding the password to a
config file or ENV variable is probably not an option.

All in all it is difficult to write good rust code :sweat_smile:

Fixes #15