Rum now observes two chains
A new observer module will handle all observations of both chains. Right
now it only contains placeholder code.
A blockchain now provides a method that returns a stream of blocks. This
stream is used by the observer to act upon new blocks.

In the vein of idiomatic rust, alot of `IntoX` and `FromX` custom traits
were replaced with the std `From` trait.

Fixes #16