Blocks stream now includes events
For each block, it now also gets the events and adds them to the
`events` field of the `Block` struct as part of the stream pipeline.

I also fixed an issue where the stream wolud yield an error. Before, the
stream would terminate after the error. Now, the error is printed, but
the stream continues polling the node.
I had the error that sometimes the block could not be retrieved when
ganache rolled back to a previous snapshot.

Ref. #18