Rum now has an events module
All the events handling was put in a separate module under the observer
namespace. Almost all code relevant to events is inside that module.
The only exceptions are the handler instantiation in the observer module
and the logs parsing call in the ethereum module.

A handler is used to convert logs into events. Based on the event
factories that are registered with the handler, logs get parsed into
events. Only logs get converted when a factory that matches log topic
and contract address is registered with the handler.

Helper methods exist in the event module to create two event handlers:
one for origin and one for auxiliary. These helper methods register the
relevant factories with their respective handlers.

Added unit tests for the events module inside the events module.

Fixes #35